feat: move filters, toast types, and creator interfaces to commons
Move the following to commons KMP module:
- FilterCommunityPosts (community post filter functions)
- FilterPostsByScopes (NIP-22 comment scope filters)
- AdditiveComplexFeedFilter (abstract feed filter)
- ToastMsg hierarchy (ToastMsg, StringToastMsg, ResourceToastMsg, ThrowableToastMsg)
- IExpiration interface
- IZapRaiser interface

Original files replaced with typealias re-exports for backward compatibility.
Fix cross-module smart cast in DisplayErrorMessages.kt.

Build verified: commons:compileKotlinJvm + amethyst:compilePlayDebugKotlin
