+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/home/datasource/nip22Comments/FilterPostsByScopes.kt
@@ -20,30 +20,16 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.home.datasource.nip22Comments
 
+// Re-export from commons for backwards compatibility
 import com.vitorpamplona.quartz.nip01Core.relay.client.pool.RelayBasedFilter
-import com.vitorpamplona.quartz.nip01Core.relay.filters.Filter
 import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import com.vitorpamplona.quartz.nip22Comments.CommentEvent
+import com.vitorpamplona.amethyst.commons.ui.screen.loggedIn.home.datasource.nip22Comments.CommentKinds as CommonsCommentKinds
+import com.vitorpamplona.amethyst.commons.ui.screen.loggedIn.home.datasource.nip22Comments.filterHomePostsByScopes as commonsFilterHomePostsByScopes
 
-val CommentKinds = listOf(CommentEvent.KIND)
+val CommentKinds = CommonsCommentKinds
 
 fun filterHomePostsByScopes(
     relay: NormalizedRelayUrl,
     scopesToLoad: Set<String>,
     since: Long?,
-): List<RelayBasedFilter> {
-    if (scopesToLoad.isEmpty()) return emptyList()
-
-    return listOf(
-        RelayBasedFilter(
-            relay = relay,
-            filter =
-                Filter(
-                    kinds = CommentKinds,
-                    tags = mapOf("I" to scopesToLoad.toList()),
-                    limit = 100,
-                    since = since,
-                ),
-        ),
-    )
-}
+): List<RelayBasedFilter> = commonsFilterHomePostsByScopes(relay, scopesToLoad, since)
